new-spec button: drop the naming modal, go straight to chat
Chat-first new-project UX. Click "+ New spec" and you land in the
seeding chat on an "Untitled" spec with an auto-generated filesystem
slug (untitled-<base36-timestamp>). The marketing phase (or a later
manual rename) names it when the product actually has a shape.

This dissolves the catch-22 we talked through: you shouldn't have to
pick a name before you've had the conversation that tells you what the
thing is. Closest analog is ChatGPT's "start typing and the title gets
generated from the first message" flow.

Removed:
- NewProjectDialog (modal asking for display name + slug)
- Its test file
No callers remain — kept intentionally unused would be cruft.

Label change: "+ New Project" → "+ New spec" to match the vocabulary
we landed on (specs are first-class, projects are specs that graduated
to being built).

Not in this PR (future):
- Rename-spec UI (inline edit in project header)
- Auto-rename when marketing phase proposes a name
- "Untitled" shown in project list gets replaced as soon as the user
  or marketing phase picks a real name

Tests: 285/285. Dashboard typechecks. Homepage renders "+ New spec".
