--- dashboard/src/components/new-project-dialog.tsx
@@ -1,111 +1,0 @@
-'use client'
-
-import { useState } from 'react'
-import { useRouter } from 'next/navigation'
-import { createBridgeProject, isBridgeEnabled } from '@/lib/bridge-client'
-
-export function NewProjectDialog({ open, onClose }: { open: boolean; onClose: () => void }) {
-  const [slug, setSlug] = useState('')
-  const [name, setName] = useState('')
-  const [slugEdited, setSlugEdited] = useState(false)
-  const [error, setError] = useState<string | null>(null)
-  const [submitting, setSubmitting] = useState(false)
-  const router = useRouter()
-
-  if (!open) return null
-
-  async function handleSubmit(e: React.FormEvent) {
-    e.preventDefault()
-    if (!isBridgeEnabled()) {
-      setError('Bridge is not enabled — cannot create projects')
-      return
-    }
-    setSubmitting(true)
-    setError(null)
-    try {
-      const result = await createBridgeProject(slug, name)
-      onClose()
-      router.push(`/projects/${result.slug}`)
-    } catch (err) {
-      setError(err instanceof Error ? err.message : String(err))
-      setSubmitting(false)
-    }
-  }
-
-  // Auto-slug from name: lowercase, spaces → hyphens
-  function handleNameChange(v: string) {
-    setName(v)
-    if (!slugEdited) {
-      setSlug(v.toLowerCase().replace(/[^a-z0-9]+/g, '-').replace(/^-|-$/g, ''))
-    }
-  }
-
-  return (
-    <div
-      className="fixed inset-0 z-50 flex items-center justify-center bg-black/40 p-4"
-      onClick={onClose}
-    >
-      <div
-        className="w-full max-w-md rounded-lg border border-gray-200 bg-white p-6 shadow-lg"
-        onClick={(e) => e.stopPropagation()}
-      >
-        <h2 className="text-lg font-semibold text-gray-900">New Project</h2>
-        <p className="mt-1 text-sm text-gray-500">
-          Create a new project and start seeding. This initializes a project directory in the Rouge.
-        </p>
-
-        <form onSubmit={handleSubmit} className="mt-4 space-y-4">
-          <div>
-            <label className="block text-xs font-medium text-gray-700 mb-1">Display name</label>
-            <input
-              type="text"
-              value={name}
-              onChange={(e) => handleNameChange(e.target.value)}
-              placeholder="MTG Oracle"
-              className="w-full rounded-md border border-gray-300 px-3 py-2 text-sm focus:border-gray-900 focus:outline-none"
-              autoFocus
-              required
-            />
-          </div>
-          <div>
-            <label className="block text-xs font-medium text-gray-700 mb-1">Slug</label>
-            <input
-              type="text"
-              value={slug}
-              onChange={(e) => { setSlug(e.target.value); setSlugEdited(true) }}
-              placeholder="mtg-oracle"
-              pattern="[a-z0-9][a-z0-9-]*"
-              className="w-full rounded-md border border-gray-300 px-3 py-2 text-sm font-mono focus:border-gray-900 focus:outline-none"
-              required
-            />
-            <p className="mt-1 text-xs text-gray-400">Lowercase, hyphens only. Used as the project directory name.</p>
-          </div>
-
-          {error && (
-            <div className="rounded-md border border-red-300 bg-red-50 px-3 py-2 text-xs text-red-800">
-              {error}
-            </div>
-          )}
-
-          <div className="flex justify-end gap-2 pt-2">
-            <button
-              type="button"
-              onClick={onClose}
-              className="rounded-md border border-gray-200 bg-white px-4 py-2 text-sm text-gray-700 hover:bg-gray-50"
-              disabled={submitting}
-            >
-              Cancel
-            </button>
-            <button
-              type="submit"
-              className="rounded-md bg-gray-900 px-4 py-2 text-sm font-medium text-white hover:bg-gray-800 disabled:opacity-50"
-              disabled={submitting || !slug || !name}
-            >
-              {submitting ? 'Creating…' : 'Create project'}
-            </button>
-          </div>
-        </form>
-      </div>
-    </div>
-  )
-}
